requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
crewai==0.16.3
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
python-multipart==0.0.9
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
twilio==8.12.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.15.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
textblob==0.17.1
langdetect==1.0.9
deepseek-ai==0.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 